Move DerivedDynamoFormat to it's own file
